+++ Sirius/src/eval/eval_state.cpp
@@ -12,8 +12,12 @@ EvalState::EvalState()
     m_CurrEntry = &m_Stack[0];
 }
 
-void EvalState::init(const Board& board)
+void EvalState::init(const Board& board, PawnTable& pawnTable)
 {
+    std::fill(m_Stack.begin(), m_Stack.end(), StackEntry{});
+    m_CurrEntry = &m_Stack[0];
+    m_PawnTable = &pawnTable;
+
     currEntry().psqtState.init();
     for (Color c : {Color::WHITE, Color::BLACK})
         for (PieceType pt : {PieceType::PAWN, PieceType::KNIGHT, PieceType::BISHOP, PieceType::ROOK, PieceType::QUEEN, PieceType::KING})
@@ -23,12 +27,18 @@ void EvalState::init(const Board& board)
                 currEntry().psqtState.addPiece(c, pt, pieces.poplsb());
         }
 
-    currEntry().pawnStructure = PawnStructure(board);
-    currEntry().passers = {};// evaluatePassers
-    currEntry().pawnShieldStorm = {};// evaluatePawnShieldStorm
-    currEntry().knightOutposts = {};// evaluateKnightOutposts
-    currEntry().bishopPawns = {};// evaluateBishopPawns
-    currEntry().rookOpen = {};// evaluateRookOpen
+    evaluatePawns(board, currEntry().pawnStructure, m_PawnTable);
+    currEntry().passers =
+        evaluatePassedPawns<Color::WHITE>(board, currEntry().pawnStructure) -
+        evaluatePassedPawns<Color::BLACK>(board, currEntry().pawnStructure);
+    currEntry().pawnShieldStorm =
+        evaluateStormShield<Color::WHITE>(board) -
+        evaluateStormShield<Color::BLACK>(board);
+    currEntry().knightOutposts =
+        evaluateKnightOutposts<Color::WHITE>(board, currEntry().pawnStructure) -
+        evaluateKnightOutposts<Color::BLACK>(board, currEntry().pawnStructure);
+    currEntry().bishopPawns = evaluateBishopPawns<Color::WHITE>(board) - evaluateBishopPawns<Color::BLACK>(board);
+    currEntry().rookOpen = evaluateRookOpen<Color::WHITE>(board) - evaluateRookOpen<Color::BLACK>(board);
 }
 
 void EvalState::push(const Board& board, const EvalUpdates& updates)
@@ -52,32 +62,38 @@ void EvalState::push(const Board& board, const EvalUpdates& updates)
     }
 
     if (updates.changedPieces.hasAny(eval_terms::pawnStructure.deps))
-        currEntry().pawnStructure = PawnStructure(board);
+        evaluatePawns(board, currEntry().pawnStructure, m_PawnTable);
     else
         currEntry().pawnStructure = oldEntry.pawnStructure;
 
     if (updates.changedPieces.hasAny(eval_terms::passers.deps))
-        currEntry().passers = {};// evaluatePassers
+        currEntry().passers =
+            evaluatePassedPawns<Color::WHITE>(board, currEntry().pawnStructure) -
+            evaluatePassedPawns<Color::BLACK>(board, currEntry().pawnStructure);
     else
         currEntry().passers = oldEntry.passers;
 
     if (updates.changedPieces.hasAny(eval_terms::pawnShieldStorm.deps))
-        currEntry().pawnShieldStorm = {};// evaluatePawnShieldStorm
+        currEntry().pawnShieldStorm =
+            evaluateStormShield<Color::WHITE>(board) -
+            evaluateStormShield<Color::BLACK>(board);
     else
         currEntry().pawnShieldStorm = oldEntry.pawnShieldStorm;
 
     if (updates.changedPieces.hasAny(eval_terms::knightOutposts.deps))
-        currEntry().knightOutposts = {};// evaluateKnightOutposts
+        currEntry().knightOutposts =
+            evaluateKnightOutposts<Color::WHITE>(board, currEntry().pawnStructure) -
+            evaluateKnightOutposts<Color::BLACK>(board, currEntry().pawnStructure);
     else
         currEntry().knightOutposts = oldEntry.knightOutposts;
 
     if (updates.changedPieces.hasAny(eval_terms::bishopPawns.deps))
-        currEntry().bishopPawns = {};// evaluateBishopPawns
+        currEntry().bishopPawns = evaluateBishopPawns<Color::WHITE>(board) - evaluateBishopPawns<Color::BLACK>(board);
     else
         currEntry().bishopPawns = oldEntry.bishopPawns;
 
     if (updates.changedPieces.hasAny(eval_terms::rookOpen.deps))
-        currEntry().rookOpen = {};// evaluateRookOpen
+        currEntry().rookOpen = evaluateRookOpen<Color::WHITE>(board) - evaluateRookOpen<Color::BLACK>(board);
     else
         currEntry().rookOpen = oldEntry.rookOpen;
 }
@@ -85,6 +101,28 @@ void EvalState::push(const Board& board, const EvalUpdates& updates)
 void EvalState::pop()
 {
     m_CurrEntry--;
+}
+
+PackedScore EvalState::score(const Board& board) const
+{
+    return
+        currEntry().psqtState.evaluate(board) +
+        currEntry().pawnStructure.score +
+        currEntry().passers +
+        currEntry().pawnShieldStorm +
+        currEntry().knightOutposts +
+        currEntry().bishopPawns +
+        currEntry().rookOpen;
+}
+
+const PawnStructure& EvalState::pawnStructure() const
+{
+    return currEntry().pawnStructure;
+}
+
+int EvalState::phase() const
+{
+    return currEntry().psqtState.phase;
 }
